+++ CHANGELOG.md
@@ -8,7 +8,13 @@
 
 ### Updates
 
+* Updated [Microsoft.Azure.DurableTask.Core](https://www.nuget.org/packages/Microsoft.Azure.DurableTask.Core) dependency to v2.10.0.
+* Fixed issue where de-dupe status values on orchestration creation were being ignored ([#120](https://github.com/microsoft/durabletask-mssql/issues/120)).
 * Fixed issue where external client ignores task hub name configuration ([#128](https://github.com/microsoft/durabletask-mssql/issues/128)) - contributed by [@bhugot](https://github.com/bhugot)
+
+### Breaking changes
+
+* Attempting to recreate a running orchestration with the same instance ID now throws `OrchestrationAlreadyExistsException` instead of `InvalidOperationException`. In most cases, this will be non-breaking since `OrchestrationAlreadyExistsException` is a sub-class of `InvalidOperationException`.
 
 ## v1.0.1